ecohome_solution/requirements.txt
# ── Runtime dependencies ────────────────────────────────────────────
# Install with: pip install -r requirements.txt

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# OpenAI client
openai>=1.40.0

# Database
sqlalchemy>=2.0.23

# Vector store
chromadb>=0.5.0

# Validation and settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
loguru>=0.7.2

# Environment
python-dotenv>=1.0.0

# Data utilities
pandas>=2.2.3
numpy>=1.26.4
python-dateutil>=2.8.2
